+++ packages/dag-designer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @robota-sdk/dag-designer
 
+## 3.0.0-beta.47
+
+### Patch Changes
+
+- @robota-sdk/dag-core@3.0.0-beta.47
+- @robota-sdk/dag-api@3.0.0-beta.47
+
 ## 3.0.0-beta.46
 
 ### Patch Changes