+++ packages/dag-nodes/llm-text-openai/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @robota-sdk/dag-node-llm-text-openai
 
+## 3.0.0-beta.47
+
+### Patch Changes
+
+- @robota-sdk/agent-core@3.0.0-beta.47
+- @robota-sdk/agent-provider-openai@3.0.0-beta.47
+- @robota-sdk/dag-core@3.0.0-beta.47
+- @robota-sdk/dag-node@3.0.0-beta.47
+
 ## 3.0.0-beta.46
 
 ### Patch Changes