+++ packages/dag-nodes/seedance-video/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @robota-sdk/dag-node-seedance-video
 
+## 3.0.0-beta.47
+
+### Patch Changes
+
+- @robota-sdk/agent-core@3.0.0-beta.47
+- @robota-sdk/agent-provider-bytedance@3.0.0-beta.47
+- @robota-sdk/dag-core@3.0.0-beta.47
+- @robota-sdk/dag-node@3.0.0-beta.47
+
 ## 3.0.0-beta.46
 
 ### Patch Changes