+++ packages/dag-worker/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @robota-sdk/dag-worker
 
+## 3.0.0-beta.47
+
+### Patch Changes
+
+- @robota-sdk/dag-core@3.0.0-beta.47
+
 ## 3.0.0-beta.46
 
 ### Patch Changes